Implement socket task framework for periodic discovery events on sockets.
1) Move socket and peer address structures out to a separate header file which defines the
   corresponding structures. Add a task structure to the socket structure to represent the
   task pqueue element of the socket.
2) Implement a new task type (TASK_TYPE_SOCKET), which is dispatched to a corresponding
   function in socket.c for later implementing the actual muticast discovery packet send.
   This function currently just reschedules the task in the configured interval.
3) Initialize the socket task when creating the socket by checking whether the bind address
   has an interval set up (i.e., non-zero and not FASTD_TIMEOUT_INV), in which case the
   socket task is immediately started. When no interval set up (for non-multicast sockets)
   or for ephemeral sockets (for an individual peer), don't set up a socket task, as it
   is useless.
4) Make sure to remove the socket task from the task queue when a socket is closed.
