+++ src/fastd.h
@@ -22,6 +22,7 @@
 #include "polling.h"
 #include "sem.h"
 #include "shell.h"
+#include "socket.h"
 #include "task.h"
 #include "util.h"
 #include "vector.h"
@@ -127,13 +128,6 @@ struct fastd_protocol {
 	bool (*describe_peer)(const fastd_peer_t *peer, char *buf, size_t len);
 };
 
-/** An union storing an IPv4 or IPv6 address */
-union fastd_peer_address {
-	struct sockaddr sa;      /**< A sockaddr field (for access to sa_family) */
-	struct sockaddr_in in;   /**< An IPv4 address */
-	struct sockaddr_in6 in6; /**< An IPv6 address */
-};
-
 #define FASTD_BIND_DEFAULT_IPV4 (1U << 1)
 #define FASTD_BIND_DEFAULT_IPV6 (1U << 2)
 #define FASTD_BIND_DYNAMIC (1U << 3)
@@ -148,15 +142,6 @@ struct fastd_bind_address {
 	fastd_timeout_t interval;        /**< Discovery interval if discovery enabled, or FASTD_TIMEOUT_INV */
 };
 
-/** A socket descriptor */
-struct fastd_socket {
-	fastd_poll_fd_t fd;               /**< The file descriptor for the socket */
-	const fastd_bind_address_t *addr; /**< The address this socket is supposed to be bound to (or NULL) */
-	fastd_peer_address_t bound_addr;  /**< The actual address that was bound to (may differ from addr when addr has
-					     a random port) and also the source address to require for incoming packets */
-	fastd_peer_t *peer; /**< If the socket belongs to a single peer (as it was create dynamically when sending a
-			       handshake), contains that peer */
-};
 
 /** A TUN/TAP interface */
 struct fastd_iface {
@@ -474,57 +459,6 @@ static inline fastd_eth_addr_t fastd_buffer_dest_address(const fastd_buffer_t *b
 	memcpy(&ret, buffer->data + offsetof(fastd_eth_header_t, dest), sizeof(fastd_eth_addr_t));
 	return ret;
 }
-
-/** Checks if a fastd_peer_address_t is the IPv4 any address */
-static inline bool fastd_peer_address_is_v4_any(const fastd_peer_address_t *addr) {
-	return addr->sa.sa_family == AF_INET && addr->in.sin_addr.s_addr == INADDR_ANY;
-}
-
-/** Checks if a fastd_peer_address_t is an IPv4 multicast address */
-static inline bool fastd_peer_address_is_v4_multicast(const fastd_peer_address_t *addr) {
-	return addr->sa.sa_family == AF_INET && IN_MULTICAST(ntohl(addr->in.sin_addr.s_addr));
-}
-
-/** Checks if host parts of two IPv4 fastd_peer_address_t are equal */
-static inline bool fastd_peer_address_is_v4_host_equal(const fastd_peer_address_t *addr1, const fastd_peer_address_t *addr2) {
-	return addr1->sa.sa_family == AF_INET && addr2->sa.sa_family == AF_INET && addr1->in.sin_addr.s_addr == addr2->in.sin_addr.s_addr;
-}
-
-/** Checks if a fastd_peer_address_t is the IPv6 any address */
-static inline bool fastd_peer_address_is_v6_any(const fastd_peer_address_t *addr) {
-	return addr->sa.sa_family == AF_INET6 && IN6_IS_ADDR_UNSPECIFIED(&addr->in6.sin6_addr);
-}
-
-/** Checks if a fastd_peer_address_t is an IPv6 link-local address */
-static inline bool fastd_peer_address_is_v6_ll(const fastd_peer_address_t *addr) {
-	return addr->sa.sa_family == AF_INET6 && IN6_IS_ADDR_LINKLOCAL(&addr->in6.sin6_addr);
-}
-
-/** Checks if a fastd_peer_address_t is an IPv6 multicast address */
-static inline bool fastd_peer_address_is_v6_multicast(const fastd_peer_address_t *addr) {
-	return addr->sa.sa_family == AF_INET6 && IN6_IS_ADDR_MULTICAST(&addr->in6.sin6_addr);
-}
-
-/** Checks if host parts of two IPv6 fastd_peer_address_t are equal */
-static inline bool fastd_peer_address_is_v6_host_equal(const fastd_peer_address_t *addr1, const fastd_peer_address_t *addr2) {
-	return addr1->sa.sa_family == AF_INET6 && addr2->sa.sa_family == AF_INET6 && IN6_ARE_ADDR_EQUAL(&addr1->in6.sin6_addr, &addr2->in6.sin6_addr);
-}
-
-/** Checks if the fastd_peer_address_t represents the IPv4 or IPv6 any address */
-static inline bool fastd_peer_address_is_any(const fastd_peer_address_t *addr) {
-	return fastd_peer_address_is_v4_any(addr) || fastd_peer_address_is_v6_any(addr);
-}
-
-/** Checks if the fastd_peer_address_t represents an IPv4 or IPv6 multicast address */
-static inline bool fastd_peer_address_is_multicast(const fastd_peer_address_t *addr) {
-	return fastd_peer_address_is_v4_multicast(addr) || fastd_peer_address_is_v6_multicast(addr);
-}
-
-/** Checks if host parts of two fastd_peer_address_t are equal */
-static inline bool fastd_peer_address_is_host_equal(const fastd_peer_address_t *addr1, const fastd_peer_address_t *addr2) {
-	return fastd_peer_address_is_v4_host_equal(addr1, addr2) || fastd_peer_address_is_v6_host_equal(addr1, addr2);
-}
-
 
 /** Duplicates a string, creating a one-element string stack */
 static inline fastd_string_stack_t *fastd_string_stack_dup(const char *str) {
